+++ src/main/java/com/leokom/chess/MainRunner.java
@@ -2,7 +2,7 @@ package com.leokom.chess;
 
 
 import com.leokom.chess.player.Player;
-import com.leokom.chess.player.legalMover.EnginePlayer;
+import com.leokom.chess.player.simple.SimpleEnginePlayer;
 import com.leokom.chess.player.winboard.WinboardPlayer;
 import org.apache.log4j.Logger;
 
@@ -20,7 +20,7 @@ public final class MainRunner {
 		logger.info( "Starting the chess..." );
 
 		final Player winboardPlayer = WinboardPlayer.create();
-		final Player enginePlayer = new EnginePlayer();
+		final Player enginePlayer = new SimpleEnginePlayer();
 		//TODO: this double setting
 		//indicates we need some master Game object
 		//that will combine them together
